Switch out the use of os.linesep for \n
This change is to fix a bug on Windows systems which showed the incorrect
output for a print statement.
In the live coding display,
	sys.stdout.write('a\n')
was shown instead of the expected
    print('a')

The fix is based on the idea that the print function's default end of line
character is \n, a newline, and not the Windows \r\n, carridge return and
newline.