Add pass only methods for screen.tracer and screen.update
Fixes #167
The Turtle Screen object has the methods tracer and update to draw the
turtle immediately on screen instead of being animated.
This does not work with the live turtle visualization created by this project.
But these two methods are something a typical turtle user could use in a script,
so they should be handled without error if called in a script.

I've added the two method as pass only methods to prevent an error, and I've
added a test verifying the tracer and update attributes exist.
Surprisingly, the pass only methods seem to be enough to satisfy the rest of the
testing suite.